+++ packages/backend-common/src/reading/GithubUrlReader.test.ts
@@ -15,20 +15,19 @@
  */
 
 import { ConfigReader } from '@backstage/config';
-import { setupServer } from 'msw/node';
+import { GitHubIntegrationConfig } from '@backstage/integration';
 import { msw } from '@backstage/test-utils';
+import fs from 'fs';
 import { rest } from 'msw';
+import { setupServer } from 'msw/node';
+import path from 'path';
 import {
   getApiRequestOptions,
   getApiUrl,
   getRawRequestOptions,
   getRawUrl,
   GithubUrlReader,
-  ProviderConfig,
-  readConfig,
 } from './GithubUrlReader';
-import fs from 'fs';
-import path from 'path';
 import { ReadTreeResponseFactory } from './tree';
 
 const treeResponseFactory = ReadTreeResponseFactory.create({
@@ -38,19 +37,19 @@ const treeResponseFactory = ReadTreeResponseFactory.create({
 describe('GithubUrlReader', () => {
   describe('getApiRequestOptions', () => {
     it('sets the correct API version', () => {
-      const config: ProviderConfig = { host: '', apiBaseUrl: '' };
+      const config: GitHubIntegrationConfig = { host: '', apiBaseUrl: '' };
       expect((getApiRequestOptions(config).headers as any).Accept).toEqual(
         'application/vnd.github.v3.raw',
       );
     });
 
     it('inserts a token when needed', () => {
-      const withToken: ProviderConfig = {
+      const withToken: GitHubIntegrationConfig = {
         host: '',
         apiBaseUrl: '',
         token: 'A',
       };
-      const withoutToken: ProviderConfig = {
+      const withoutToken: GitHubIntegrationConfig = {
         host: '',
         apiBaseUrl: '',
       };
@@ -65,12 +64,12 @@ describe('GithubUrlReader', () => {
 
   describe('getRawRequestOptions', () => {
     it('inserts a token when needed', () => {
-      const withToken: ProviderConfig = {
+      const withToken: GitHubIntegrationConfig = {
         host: '',
         rawBaseUrl: '',
         token: 'A',
       };
-      const withoutToken: ProviderConfig = {
+      const withoutToken: GitHubIntegrationConfig = {
         host: '',
         rawBaseUrl: '',
       };
@@ -85,12 +84,12 @@ describe('GithubUrlReader', () => {
 
   describe('getApiUrl', () => {
     it('rejects targets that do not look like URLs', () => {
-      const config: ProviderConfig = { host: '', apiBaseUrl: '' };
+      const config: GitHubIntegrationConfig = { host: '', apiBaseUrl: '' };
       expect(() => getApiUrl('a/b', config)).toThrow(/Incorrect URL: a\/b/);
     });
 
     it('happy path for github', () => {
-      const config: ProviderConfig = {
+      const config: GitHubIntegrationConfig = {
         host: 'github.com',
         apiBaseUrl: 'https://api.github.com',
       };
@@ -117,7 +116,7 @@ describe('GithubUrlReader', () => {
     });
 
     it('happy path for ghe', () => {
-      const config: ProviderConfig = {
+      const config: GitHubIntegrationConfig = {
         host: 'ghe.mycompany.net',
         apiBaseUrl: 'https://ghe.mycompany.net/api/v3',
       };
@@ -136,12 +135,12 @@ describe('GithubUrlReader', () => {
 
   describe('getRawUrl', () => {
     it('rejects targets that do not look like URLs', () => {
-      const config: ProviderConfig = { host: '', apiBaseUrl: '' };
+      const config: GitHubIntegrationConfig = { host: '', apiBaseUrl: '' };
       expect(() => getRawUrl('a/b', config)).toThrow(/Incorrect URL: a\/b/);
     });
 
     it('happy path for github', () => {
-      const config: ProviderConfig = {
+      const config: GitHubIntegrationConfig = {
         host: 'github.com',
         rawBaseUrl: 'https://raw.githubusercontent.com',
       };
@@ -158,7 +157,7 @@ describe('GithubUrlReader', () => {
     });
 
     it('happy path for ghe', () => {
-      const config: ProviderConfig = {
+      const config: GitHubIntegrationConfig = {
         host: 'ghe.mycompany.net',
         rawBaseUrl: 'https://ghe.mycompany.net/raw',
       };
@@ -170,60 +169,6 @@ describe('GithubUrlReader', () => {
       ).toEqual(
         new URL('https://ghe.mycompany.net/raw/a/b/branchname/path/to/c.yaml'),
       );
-    });
-  });
-
-  describe('readConfig', () => {
-    function config(
-      providers: { host: string; apiBaseUrl?: string; token?: string }[],
-    ) {
-      return ConfigReader.fromConfigs([
-        {
-          context: '',
-          data: {
-            integrations: { github: providers },
-          },
-        },
-      ]);
-    }
-
-    it('adds a default GitHub entry when missing', () => {
-      const output = readConfig(config([]));
-      expect(output).toEqual([
-        {
-          host: 'github.com',
-          apiBaseUrl: 'https://api.github.com',
-          rawBaseUrl: 'https://raw.githubusercontent.com',
-        },
-      ]);
-    });
-
-    it('injects the correct GitHub API base URL when missing', () => {
-      const output = readConfig(config([{ host: 'github.com' }]));
-      expect(output).toEqual([
-        {
-          host: 'github.com',
-          apiBaseUrl: 'https://api.github.com',
-          rawBaseUrl: 'https://raw.githubusercontent.com',
-        },
-      ]);
-    });
-
-    it('rejects custom targets with no base URLs', () => {
-      expect(() => readConfig(config([{ host: 'ghe.company.com' }]))).toThrow(
-        "GitHub integration for 'ghe.company.com' must configure an explicit apiBaseUrl and rawBaseUrl",
-      );
-    });
-
-    it('rejects funky configs', () => {
-      expect(() => readConfig(config([{ host: 7 } as any]))).toThrow(/host/);
-      expect(() => readConfig(config([{ token: 7 } as any]))).toThrow(/token/);
-      expect(() =>
-        readConfig(config([{ host: 'github.com', apiBaseUrl: 7 } as any])),
-      ).toThrow(/apiBaseUrl/);
-      expect(() =>
-        readConfig(config([{ host: 'github.com', token: 7 } as any])),
-      ).toThrow(/token/);
     });
   });
 
@@ -271,6 +216,7 @@ describe('GithubUrlReader', () => {
       const processor = new GithubUrlReader(
         {
           host: 'github.com',
+          apiBaseUrl: 'https://api.github.com',
         },
         { treeResponseFactory },
       );
@@ -293,6 +239,7 @@ describe('GithubUrlReader', () => {
       const processor = new GithubUrlReader(
         {
           host: 'github.com',
+          apiBaseUrl: 'https://api.github.com',
         },
         { treeResponseFactory },
       );
@@ -308,6 +255,7 @@ describe('GithubUrlReader', () => {
       const processor = new GithubUrlReader(
         {
           host: 'github.com',
+          apiBaseUrl: 'https://api.github.com',
         },
         { treeResponseFactory },
       );