+++ packages/core/src/api-wrappers/createApp.tsx
@@ -61,7 +61,11 @@ export const defaultConfigLoader: AppConfigLoader = async (
   if (runtimeConfigJson !== '__app_injected_runtime_config__'.toUpperCase()) {
     try {
       const data = JSON.parse(runtimeConfigJson) as JsonObject;
-      configs.push({ data, context: 'env' });
+      if (Array.isArray(data)) {
+        configs.push(...data);
+      } else {
+        configs.push({ data, context: 'env' });
+      }
     } catch (error) {
       throw new Error(`Failed to load runtime configuration, ${error}`);
     }