+++ packages/integration/src/gitlab/index.ts
@@ -14,14 +14,8 @@
  * limitations under the License.
  */
 
-import { Command } from 'commander';
-import { buildBundle } from '../../lib/bundler';
-import { loadCliConfig } from '../../lib/config';
-
-export default async (cmd: Command) => {
-  await buildBundle({
-    entry: 'dev/index',
-    statsJsonEnabled: cmd.stats,
-    ...(await loadCliConfig(cmd.config)),
-  });
-};
+export {
+  readGitLabIntegrationConfig,
+  readGitLabIntegrationConfigs,
+} from './config';
+export type { GitLabIntegrationConfig } from './config';