+++ plugins/catalog-graphql/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-catalog-graphql
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies [1722cb53c]
+- Updated dependencies [1722cb53c]
+- Updated dependencies [7b37e6834]
+- Updated dependencies [8e2effb53]
+  - @backstage/backend-common@0.3.0
+
 ## 0.2.0
 
 ### Minor Changes